dictquick handles all button stuff
